+++ vllm/models/minimax_m3/common/sparse_attention.py
@@ -20,7 +20,7 @@ from typing import ClassVar
 
 import torch
 
-from vllm import envs
+from vllm._aiter_ops import rocm_aiter_ops
 from vllm.config import VllmConfig
 from vllm.config.cache import CacheDType
 from vllm.forward_context import get_forward_context
@@ -61,10 +61,9 @@ logger = init_logger(__name__)
 
 def minimax_m3_use_aiter_sparse_pa(num_kv_heads: int) -> bool:
     """Whether to use the ROCm AITER page-16 sparse PA prototype."""
-    return bool(
-        current_platform.is_rocm()
-        and envs.VLLM_ROCM_USE_AITER
-        and envs.VLLM_ROCM_SHUFFLE_KV_CACHE_LAYOUT
+    return (
+        rocm_aiter_ops.is_enabled()
+        and rocm_aiter_ops.is_shuffle_kv_cache_enabled()
         and num_kv_heads == 1
     )
 
@@ -127,7 +126,7 @@ class MiniMaxM3SparseBackend(AttentionBackend):
         # layout we want.
         if include_num_layers_dimension:
             raise NotImplementedError  # no cross-layer KV blocks in M3
-        if envs.VLLM_ROCM_USE_AITER and envs.VLLM_ROCM_SHUFFLE_KV_CACHE_LAYOUT:
+        if rocm_aiter_ops.is_enabled() and rocm_aiter_ops.is_shuffle_kv_cache_enabled():
             # The AITER page-16 sparse PA path reinterprets the K and V slices
             # as separate SHUFFLE caches. Keep K/V physically separated so
             # kv_cache[:, 0] and kv_cache[:, 1] are contiguous byte ranges.
@@ -424,86 +423,6 @@ class MiniMaxM3SparseTritonImpl(MiniMaxM3SparseImpl):
         return output
 
 
-class MiniMaxM3SparseAiterPAImpl(MiniMaxM3SparseImpl):
-    """ROCm AITER page-16 SHUFFLE sparse paged attention."""
-
-    def forward(
-        self,
-        layer: AttentionLayer,
-        query: torch.Tensor,
-        kv_cache: torch.Tensor,
-        output: torch.Tensor,
-    ) -> torch.Tensor:
-        from vllm.models.minimax_m3.amd.ops.sparse_pa import (
-            minimax_m3_sparse_attn_decode_aiter,
-            minimax_m3_sparse_attn_prefill_aiter,
-        )
-
-        attn_metadata = get_forward_context().attn_metadata
-        if not isinstance(attn_metadata, dict):
-            return output
-        main_md = attn_metadata[layer.layer_name]  # type: ignore[attr-defined]
-        assert isinstance(main_md, MiniMaxM3SparseMetadata)
-
-        nd = main_md.num_decode_tokens
-        num_tokens = main_md.num_actual_tokens
-        topk = layer.topk_indices_buffer  # type: ignore[attr-defined]
-        assert topk is not None
-        if self.num_kv_heads != 1:
-            raise NotImplementedError(
-                "MiniMax-M3 AITER sparse PA currently requires per-rank "
-                f"num_kv_heads == 1, got {self.num_kv_heads}"
-            )
-
-        hd = self.head_size
-        q = query[:num_tokens].view(-1, self.num_heads, hd)
-        out = output[:num_tokens].view(-1, self.num_heads, hd)
-        k_cache, v_cache = layer.get_aiter_sparse_pa_kv_cache()  # type: ignore[attr-defined]
-        k_scale = getattr(layer, "_k_scale", None) if self.use_fp8_kv else None
-        v_scale = getattr(layer, "_v_scale", None) if self.use_fp8_kv else None
-
-        if main_md.num_decodes > 0:
-            d = main_md.decode
-            assert d is not None
-            if d.decode_query_len != 1:
-                raise NotImplementedError(
-                    "MiniMax-M3 AITER sparse PA does not support speculative "
-                    f"decode_query_len={d.decode_query_len}"
-                )
-            minimax_m3_sparse_attn_decode_aiter(
-                q[:nd],
-                k_cache,
-                v_cache,
-                topk[:, :nd, :],
-                d.block_table,
-                d.seq_lens,
-                self.num_kv_heads,
-                self.scale,
-                out[:nd],
-                k_scale=k_scale,
-                v_scale=v_scale,
-            )
-
-        if main_md.num_prefills > 0:
-            p = main_md.prefill
-            assert p is not None
-            minimax_m3_sparse_attn_prefill_aiter(
-                q[nd:],
-                k_cache,
-                v_cache,
-                topk[:, nd:num_tokens, :],
-                p.block_table,
-                p.cu_seqlens_q,
-                p.context_lens,
-                self.num_kv_heads,
-                self.scale,
-                out[nd:],
-                k_scale=k_scale,
-                v_scale=v_scale,
-            )
-        return output
-
-
 def select_main_impl_cls(
     *,
     topk_blocks: int,
@@ -513,9 +432,10 @@ def select_main_impl_cls(
     """Pick the main attend impl off the main KV-cache dtype.
 
     Blackwell (SM100) uses the MSA attend for supported top-k block counts
-    when the KV cache is BF16 or FP8 E4M3; non-Blackwell and FP8 E5M2 fall
-    back to Triton. The MSA module is imported lazily so AMD/non-SM100 never
-    import fmha_sm100.
+    when the KV cache is BF16 or FP8 E4M3; MI355 uses AITER sparse PA
+    with shuffle KV cache layout; Other platforms and FP8 E5M2 fall
+    back to Triton. The MSA modules are imported lazily avoid import errors
+    on unsupported platforms.
     """
     use_aiter_sparse_pa = minimax_m3_use_aiter_sparse_pa(num_kv_heads)
     use_msa = (
@@ -534,6 +454,10 @@ def select_main_impl_cls(
         topk_blocks,
     )
     if use_aiter_sparse_pa:
+        from vllm.models.minimax_m3.amd.sparse_attention_msa import (
+            MiniMaxM3SparseAiterPAImpl,
+        )
+
         return MiniMaxM3SparseAiterPAImpl
     if use_msa:
         from vllm.models.minimax_m3.nvidia.sparse_attention_msa import (
